bridge/requirements.txt
coinbase-advanced-py
fastapi
uvicorn
python-dotenv
